requirements.txt
# Core
anthropic==0.40.0
supabase==2.9.1
python-dotenv==1.0.1

# Data
pandas==2.2.3
pandas_ta==0.3.14b
numpy==2.1.3
requests==2.32.3
httpx==0.27.2
massive>=1.0.0          # Polygon.io Python client rebranded — pip install -U massive

# Backtesting
backtesting==0.3.3
vectorbt==0.26.2

# Optimization
optuna==3.6.1
scikit-learn==1.5.2
xgboost==2.1.1

# Web / API server
fastapi==0.115.0
uvicorn[standard]==0.32.0
jinja2==3.1.4
python-multipart==0.0.12   # required for FastAPI form parsing

# Scheduling & async
apscheduler==3.10.4
asyncio-throttle==1.0.2

# Modal (cloud compute)
modal==0.64.0

# Storage
boto3==1.35.0
cloudpathlib[s3]==0.19.0

# Stats
scipy==1.14.1
statsmodels==0.14.4

# Utilities
python-dateutil==2.9.0
pytz==2024.2
tqdm==4.67.1
structlog==24.4.0
